requirements.txt
camel-ai
google-generativeai 
googlemaps
pyyaml
gps3
python-dotenv